Proof of concept register function - update 5

+++ lib/register_third.dart
@@ -9,7 +9,6 @@ import 'connection_handler.dart';
 import 'constants.dart';
 
 import 'package:flutter/cupertino.dart';
-
 class RegisterThird extends StatefulWidget {
 
   final Connection connection;
@@ -129,7 +128,7 @@ class _RegisterThirdState extends State<RegisterThird> {
                     cursorColor: Colors.green,
                     style: TextStyle(fontWeight: FontWeight.bold),
                     onChanged: (val) {
-                      setState(() => p.height= double.parse(val));
+                      setState(() => p.height = int.parse(val));
                     }
                 ),
                 SizedBox(height: 15.0),
@@ -139,7 +138,7 @@ class _RegisterThirdState extends State<RegisterThird> {
                     cursorColor: Colors.green,
                     style: TextStyle(fontWeight: FontWeight.bold),
                     onChanged: (val) {
-                      setState(() => p.weight = double.parse(val));
+                      setState(() => p.weight = int.parse(val));
                     }
                 ),
                 SizedBox(height: 150),
